added sub component handling auth input and status rendering
