+++ src/components/hiveAuthModal/container/hiveAuthModal.tsx
@@ -1,32 +1,23 @@
 import React, { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react';
-import { Linking, View, Text, ActivityIndicator, Keyboard } from 'react-native';
+import { Linking, View, Keyboard } from 'react-native';
 
 import HAS from 'hive-auth-wrapper';
 import { v4 as uuidv4 } from 'uuid';
 import { useIntl } from 'react-intl';
 import { HiveSignerMessage } from 'utils/hive-signer-helper';
-import { debounce } from 'lodash';
 import { useNavigation } from '@react-navigation/native';
 import ActionSheet from 'react-native-actions-sheet';
-import EStyleSheet from 'react-native-extended-stylesheet';
-import Animated, {
-  FadeOut,
-  LinearTransition,
-  FadeInUp,
-  FadeOutDown,
-  ZoomIn,
-  ZoomOut,
-} from 'react-native-reanimated';
 import styles from '../styles/hiveAuthModal.styles';
-import { getDigitPinCode, lookupAccounts } from '../../../providers/hive/dhive';
+import { getDigitPinCode } from '../../../providers/hive/dhive';
 import { loginWithHiveAuth } from '../../../providers/hive/auth';
 import { useAppSelector, usePostLoginActions } from '../../../hooks';
 import ROUTES from '../../../constants/routeNames';
 import AUTH_TYPE from '../../../constants/authType';
 import { decryptKey } from '../../../utils/crypto';
-import { FormInput, Icon, MainButton, ModalHeader } from '../..';
-import HIVE_AUTH_ICON from '../../../assets/HiveAuth_logo.png';
+import { ModalHeader } from '../..';
 import { delay } from '../../../utils/editor';
+import { AuthInputContent } from '../children/authInputContent';
+import { StatusContent } from '../children/statusContent';
 
 const APP_META = {
   name: 'Ecency',
@@ -34,13 +25,8 @@ const APP_META = {
   icon: undefined,
 };
 
-enum Modes {
-  AUTH = 0,
-  SIGN = 1,
-}
-
-enum Status {
-  IDLE = 0,
+export enum HiveAuthStatus {
+  INPUT = 0,
   PROCESSING = 1,
   SUCCESS = 2,
   ERROR = 3,
@@ -51,28 +37,6 @@ interface HiveAuthModalProps {
 }
 
 export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) => {
-  useImperativeHandle(ref, () => ({
-    showModal: (_username: string) => {
-      if (_username) {
-        setMode(Modes.AUTH);
-        setUsername(_username || '');
-        setStatus(Status.IDLE);
-        setStatusText('');
-      }
-      bottomSheetModalRef.current?.show();
-    },
-    broadcastActiveOps: (opsArray: any) => {
-      if (opsArray) {
-        setMode(Modes.SIGN);
-        setStatus(Status.PROCESSING);
-        setStatusText('Initializing...');
-        bottomSheetModalRef.current?.show();
-        delay(2000).then(() => {
-          _broadcastActiveOps(opsArray);
-        });
-      }
-    },
-  }));
 
   const intl = useIntl();
   const navigation = useNavigation();
@@ -84,12 +48,29 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
 
   const bottomSheetModalRef = useRef();
 
-  const [mode, setMode] = useState(Modes.AUTH);
-
-  const [username, setUsername] = useState('');
+  const [initUsername, setInitUsername] = useState<string>();
   const [statusText, setStatusText] = useState('');
-  const [isUsernameValid, setIsUsernameValid] = useState(false);
-  const [status, setStatus] = useState(Status.IDLE);
+  const [status, setStatus] = useState(HiveAuthStatus.INPUT);
+
+  useImperativeHandle(ref, () => ({
+    showModal: (_username?: string) => {
+      setInitUsername(_username);
+      setStatus(HiveAuthStatus.INPUT);
+      setStatusText('');
+      bottomSheetModalRef.current?.show();
+    },
+    broadcastActiveOps: (opsArray: any) => {
+      if (opsArray) {
+
+        setStatus(HiveAuthStatus.PROCESSING);
+        setStatusText('Initializing...');
+        bottomSheetModalRef.current?.show();
+        delay(2000).then(() => {
+          _broadcastActiveOps(opsArray);
+        });
+      }
+    },
+  }));
 
   useEffect(() => {
     // Your application information
@@ -103,18 +84,11 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
     // console.log(status)
   }, []);
 
-  const debouncedCheckValidity = debounce((uname) => {
-    _checkUsernameIsValid(uname);
-  }, 500);
 
-  useEffect(() => {
-    debouncedCheckValidity(username);
-    return () => debouncedCheckValidity.cancel();
-  }, [username]);
 
-  const _sendAuthReq = async () => {
+  const handleAuthRequest = async (username:string) => {
     setStatusText('Initiating...');
-    setStatus(Status.PROCESSING);
+    setStatus(HiveAuthStatus.PROCESSING);
 
     Keyboard.dismiss();
     await delay(1000); // TODO: delay to assist modal settle height before show redirecting dialog
@@ -169,7 +143,7 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
         } else {
           // TOOD: prompt to install valid keychain app
           setStatusText('Hive Authentication App Not Installed');
-          setStatus(Status.ERROR);
+          setStatus(HiveAuthStatus.ERROR);
         }
       };
 
@@ -186,7 +160,7 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
       postLoginActions.updateAccountsData(accountData);
 
       setStatusText('Authentication Success!');
-      setStatus(Status.SUCCESS);
+      setStatus(HiveAuthStatus.SUCCESS);
 
       await delay(2000);
 
@@ -207,13 +181,13 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
     } catch (error) {
       console.warn('Login failed');
       setStatusText(error.message || 'Authentication Failed');
-      setStatus(Status.ERROR);
+      setStatus(HiveAuthStatus.ERROR);
     }
   };
 
   const _broadcastActiveOps = async (opsArray: any) => {
     try {
-      setStatus(Status.PROCESSING);
+      setStatus(HiveAuthStatus.PROCESSING);
       setStatusText('Initializing...');
       if (currentAccount.local.authType !== AUTH_TYPE.HIVE_AUTH) {
         throw new Error('Invalid auth type');
@@ -258,13 +232,13 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
         // respond back to transfer screen
       }
 
-      setStatus(Status.SUCCESS);
+      setStatus(HiveAuthStatus.SUCCESS);
       setStatusText('Transaction Successful!');
       await delay(2000);
 
       _closeModal();
     } catch (error) {
-      setStatus(Status.ERROR);
+      setStatus(HiveAuthStatus.ERROR);
       setStatusText(error.message || 'Transaction Rejected!');
     }
   };
@@ -276,120 +250,28 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
     }
   };
 
-  const _handleUsernameChange = (username: string) => {
-    const formattedUsername = username.trim().toLowerCase();
-    setUsername(formattedUsername);
-  };
 
-  const _checkUsernameIsValid = async (uname) => {
-    try {
-      const accts = await lookupAccounts(uname);
-      const isValid = accts.includes(uname);
-      setIsUsernameValid(isValid);
-    } catch (err) {
-      setIsUsernameValid(false);
-    }
-  };
+  const _renderContent = () => {
 
-  const _renderAuthContent = (
-    <Animated.View
-      style={{ width: '100%', justifyContent: 'center', alignItems: 'center' }}
-      exiting={FadeOut}
-    >
-      <Animated.View style={{ width: '90%' }} layout={LinearTransition}>
-        <FormInput
-          rightIconName="at"
-          leftIconName="close"
-          iconType="MaterialCommunityIcons"
-          isValid={!username || isUsernameValid}
-          onChange={_handleUsernameChange}
-          placeholder={intl.formatMessage({
-            id: 'login.username',
-          })}
-          isEditable
-          type="username"
-          isFirstImage
-          value={username}
-          inputStyle={styles.input}
-          wrapperStyle={styles.inputWrapper}
-          onBlur={() => _checkUsernameIsValid(username)}
-        />
-      </Animated.View>
+    const _content = status === HiveAuthStatus.INPUT
+      ? <AuthInputContent initUsername={initUsername} handleAuthRequest={handleAuthRequest} />
+      : <StatusContent status={status} statusText={statusText} />
 
-      {isUsernameValid && (
-        <Animated.View entering={ZoomIn} exiting={ZoomOut}>
-          <MainButton
-            text={intl.formatMessage({ id: 'login.signin_with' })}
-            secondText=" Hive"
-            secondTextStyle={{ color: EStyleSheet.value('$primaryRed') }}
-            style={{
-              backgroundColor: 'transparent',
-              marginTop: 12,
-              borderWidth: 1,
-              borderColor: EStyleSheet.value('$iconColor'),
-              height: 44,
-            }}
-            onPress={_sendAuthReq}
-            isLoading={status === Status.PROCESSING}
-            source={HIVE_AUTH_ICON}
-          />
-        </Animated.View>
-      )}
-    </Animated.View>
-  );
 
-  const _renderResultIcon = (iconName: string, colorId: string) => (
-    <Animated.View entering={ZoomIn.springify().duration(500)} exiting={ZoomOut}>
-      <Icon
-        iconType="AntDesign"
-        name={iconName}
-        color={EStyleSheet.value(colorId)}
-        size={88}
-        style={styles.resultIcon}
-      />
-    </Animated.View>
-  );
+    return (
+      <View style={styles.container}>
+        <ModalHeader
+          title="Hive Auth"
+          isCloseButton={true}
+          onClosePress={_closeModal} />
 
-  const _renderProcessContent = (
-    <>
-      {status === Status.SUCCESS && _renderResultIcon('checkcircleo', '$primaryGreen')}
-      {status === Status.ERROR && _renderResultIcon('closecircleo', '$primaryRed')}
+        <View style={styles.content}>
+          {_content}
+        </View>
 
-      <Animated.View
-        style={{ flexDirection: 'row', alignItems: 'center' }}
-        layout={LinearTransition}
-        entering={FadeInUp}
-        exiting={FadeOutDown}
-      >
-        {status === Status.PROCESSING && (
-          <ActivityIndicator
-            style={{ marginRight: 16 }}
-            size="large"
-            color={EStyleSheet.value('$primaryBlue')}
-          />
-        )}
-        <Text
-          style={{
-            color: EStyleSheet.value('$primaryDarkText'),
-            fontWeight: 300,
-            fontSize: 24,
-            textAlign: 'center',
-          }}
-        >
-          {statusText}
-        </Text>
-      </Animated.View>
-    </>
-  );
-
-  const _renderContent = (
-    <View style={styles.container}>
-      <ModalHeader title="Hive Auth" isCloseButton={true} onClosePress={_closeModal} />
-      <View style={styles.content}>
-        {status !== Status.IDLE ? _renderProcessContent : mode === Modes.AUTH && _renderAuthContent}
       </View>
-    </View>
-  );
+    )
+  };
 
   return (
     <ActionSheet
@@ -397,14 +279,14 @@ export const HiveAuthModal = forwardRef(({ onClose }: HiveAuthModalProps, ref) =
       gestureEnabled={false}
       hideUnderlay={true}
       onClose={() => {
-        setStatus(Status.IDLE);
+        setStatus(HiveAuthStatus.INPUT);
         setStatusText('');
-        setUsername('');
+        setInitUsername(undefined);
       }}
       containerStyle={styles.sheetContent}
       indicatorStyle={styles.indicator}
     >
-      {_renderContent}
+      {_renderContent()}
     </ActionSheet>
   );
 });
